refactor http parts
