+++ socket.h
@@ -18,6 +18,15 @@
 int socket_init(void);
 void socket_exit(void);
 
+struct camblet_socket;
+typedef struct camblet_socket camblet_socket;
+
+char *get_write_buffer(camblet_socket *s);
+// returns a pointer to the buffer where the caller can write len long data (possibly resizing the buffer)
+char *get_write_buffer_for_write(camblet_socket *s, int len); 
+int get_write_buffer_size(camblet_socket *s);
+void set_write_buffer_size(camblet_socket *s, int size);
+
 typedef enum
 {
     INPUT,
